+++ CHANGELOG.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) once
 we hit release version 1.0.0.
 
+## [0.15.2] - YYYY-MM-DD
 
-## [0.15.1] - YYYY-MM-DD
+### Fixed
+- fixed warning raised when reading electronic structure without Ef, #826
+
+
+## [0.15.1] - 2024-09-10
 
 ### Added
 - enabled `lowdin` to return the Lowdin transformation matrix, and also